Move StatusControl for Login, add error in Card
