+++ packages/ui/src/index.ts
@@ -28,7 +28,7 @@ export { getCurrentLocation, navigate, location } from './location'
 export { default as EditBox } from './components/EditBox.svelte'
 export { default as Label } from './components/Label.svelte'
 export { default as Button } from './components/Button.svelte'
-export { default as StatusControl } from './components/StatusControl.svelte'
+export { default as Status } from './components/Status.svelte'
 export { default as Component } from './components/Component.svelte'
 export { default as Icon } from './components/Icon.svelte'
 export { default as ActionIcon } from './components/ActionIcon.svelte'
